chore: move abi encoding to kms-grpc
