+++ core/grpc/Cargo.toml
@@ -12,6 +12,7 @@ path = "src/lib.rs"
 crate-type = ["lib", "cdylib"]
 
 [dependencies]
+alloy-dyn-abi = { workspace = true }
 alloy-primitives = { workspace = true, features = ["k256"] }
 alloy-sol-types = { workspace = true, features = ["eip712-serde"] }
 anyhow.workspace = true
@@ -32,9 +33,14 @@ tfhe = { workspace = true, features = [
     "experimental-force_fft_algo_dif4",
 ] }
 tfhe-versionable.workspace = true
-threshold-fhe = { workspace = true, default-features = false, features = ["extension_degree_4"] }
+threshold-fhe = { workspace = true, default-features = false, features = [
+    "extension_degree_4",
+] }
 tracing = { workspace = true, features = ["log"] }
 wasm-bindgen = { version = "=0.2.100", features = ["serde-serialize"] }
+
+[dev-dependencies]
+tracing-test = "=0.2.5"
 
 [build-dependencies]
 tonic-build.workspace = true
