+++ core/grpc/src/lib.rs
@@ -31,3 +31,20 @@ pub mod utils;
 
 // Re-export identifier types for easier access
 pub use identifiers::{IdentifierError, KeyId, RequestId};
+
+use anyhow::anyhow;
+use std::{fmt, panic::Location};
+
+// NOTE: the below is copied from core/threshold
+// since the calling tracing from another crate
+// does not generate correct logs in tracing_test::traced_test
+#[track_caller]
+pub(crate) fn anyhow_error_and_log<S: AsRef<str> + fmt::Display>(msg: S) -> anyhow::Error {
+    tracing::error!("Error in {}: {}", Location::caller(), msg);
+    anyhow_tracked(msg)
+}
+
+#[track_caller]
+pub(crate) fn anyhow_tracked<S: AsRef<str> + fmt::Display>(msg: S) -> anyhow::Error {
+    anyhow!("Error in {}: {}", Location::caller(), msg)
+}
